authentication
made the AuthContext to manage the logged in session for the user. also rendered the header to show the initial and name. made it so you can’t access most of the pages if you’re not logged in. if you’re not logged in and try, it redirects to login page. if you’re logged in and try to go to checkout, it displays a message to go through wih the booking process from the home page. also rendered dynamically the payment cards